python-ocr/requirements.txt
Pillow
pytesseract
opencv-python
numpy
googletrans==4.0.0-rc1